% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auntie-package.R
\docType{package}
\name{auntie-package}
\alias{auntie-package}
\alias{auntie}
\alias{auntieEdna}
\title{auntie-package}
\description{
Provides added tools for working within eDNA projects.
}
\section{Guide to functions}{

  See \href{https://github.com/BigelowLab/auntie}{auntie} README for examples.
}

