% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{bbmap_subsample}
\alias{bbmap_subsample}
\title{A wrapper around bbmap's \href{https://github.com/BioInfoTools/BBMap/blob/a9ceda047a7c918dc090de0fdbf6f924292d4a1f/sh/reformat.sh}{reformat.sh} to subsample.}
\usage{
bbmap_subsample(
  x,
  path = ".",
  app = which_reformat.sh(),
  reformat_args = "ow=f samplereadstarget=15000 sampleseed=7",
  overwrite = FALSE,
  stdout = ""
)
}
\arguments{
\item{x}{list or character vector.  If a list then 2 elements which are forward and reverse filenames}

\item{path}{character, the default path to write the subsampled files to}

\item{app}{character, the path tot he reformat.sh executable}

\item{reformat_args}{character, default for subsampling, see \href{https://github.com/BioInfoTools/BBMap/blob/a9ceda047a7c918dc090de0fdbf6f924292d4a1f/sh/reformat.sh}{reformat.sh}}

\item{stdout}{see \code{system2} - but really for future use if we can figure out how to capture java output - ignore for now}
}
\value{
numeric, 0 for success for all inputs
}
\description{
A wrapper around bbmap's \href{https://github.com/BioInfoTools/BBMap/blob/a9ceda047a7c918dc090de0fdbf6f924292d4a1f/sh/reformat.sh}{reformat.sh} to subsample.
}
\examples{
\dontrun{
$ module use /mod/scgc
$ module load anaconda3/2019.07
$ module load bbmap
$ module load dada2
$ R
> filepairs = auntie::list_filepairs("/mnt/storage/data/edna/mothur/projects/subsample/raw")
> sub_path <- "/mnt/storage/data/edna/mothur/projects/subsample/sub"
> charlier::make_path("/mnt/storage/data/edna/mothur/projects/subsample/sub")
> ok <- auntie::bbmap_subsample(filepairs, path = sub_path, reformat_args = "ow=t")
}
}
