% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{fastq_ids}
\alias{fastq_ids}
\title{Retrieve IDs for FastQ pairs}
\usage{
fastq_ids(x)
}
\arguments{
\item{x}{list of filepairs (forward and reverse elements)}
}
\value{
two element list of IDs
}
\description{
Retrieve IDs for FastQ pairs
}
\examples{
\dontrun{
library(dplyr)
 filepairs = dadautils::list_filepairs("/mnt/storage/data/edna/mothur/projects/subsample/sub")
 z = dplyr::as_tibble(fastq_ids(filepairs)[[1]])
 head(z)
 # # A tibble: 6 × 2
 #   forward                                                       reverse                                                      
 #   <chr>                                                         <chr>                                                        
 # 1 A01346:18:H7KY5DRXY:2:2101:25229:2002 1:N:0:TGGTTCGA+AACCGTGT A01346:18:H7KY5DRXY:2:2101:25229:2002 2:N:0:TGGTTCGA+AACCGTGT
 # 2 A01346:18:H7KY5DRXY:2:2101:10673:3004 1:N:0:TGGTTCGA+AACCGTGT A01346:18:H7KY5DRXY:2:2101:10673:3004 2:N:0:TGGTTCGA+AACCGTGT
 # 3 A01346:18:H7KY5DRXY:2:2101:2211:3223 1:N:0:TGGTTCGA+AACCGTGT  A01346:18:H7KY5DRXY:2:2101:2211:3223 2:N:0:TGGTTCGA+AACCGTGT 
 # 4 A01346:18:H7KY5DRXY:2:2101:7907:3724 1:N:0:TGGTTCAA+AACCGTGT  A01346:18:H7KY5DRXY:2:2101:7907:3724 2:N:0:TGGTTCAA+AACCGTGT 
 # 5 A01346:18:H7KY5DRXY:2:2101:15248:4820 1:N:0:TGGTTCGA+AACCGTGT A01346:18:H7KY5DRXY:2:2101:15248:4820 2:N:0:TGGTTCGA+AACCGTGT
 # 6 A01346:18:H7KY5DRXY:2:2101:28004:5368 1:N:0:TGGTTCGA+AACCGTGT A01346:18:H7KY5DRXY:2:2101:28004:5368 2:N:0:TGGTTCGA+AACCGTGT
 tail(z)
 # # A tibble: 6 × 2
 #   forward                                                        reverse                                                       
 #   <chr>                                                          <chr>                                                         
 # 1 A01346:18:H7KY5DRXY:2:2278:20826:35305 1:N:0:TGGTTCGA+TACCGTGT A01346:18:H7KY5DRXY:2:2278:20826:35305 2:N:0:TGGTTCGA+TACCGTGT
 # 2 A01346:18:H7KY5DRXY:2:2278:5222:35869 1:N:0:TGGTTCGA+TACCGTGT  A01346:18:H7KY5DRXY:2:2278:5222:35869 2:N:0:TGGTTCGA+TACCGTGT 
 # 3 A01346:18:H7KY5DRXY:2:2278:12780:36151 1:N:0:TGGTTCGA+TACCGTGT A01346:18:H7KY5DRXY:2:2278:12780:36151 2:N:0:TGGTTCGA+TACCGTGT
 # 4 A01346:18:H7KY5DRXY:2:2278:26964:36385 1:N:0:TGGTTCGA+TACCGTGT A01346:18:H7KY5DRXY:2:2278:26964:36385 2:N:0:TGGTTCGA+TACCGTGT
 # 5 A01346:18:H7KY5DRXY:2:2278:2103:36417 1:N:0:TGGTTCGA+TACCGTGT  A01346:18:H7KY5DRXY:2:2278:2103:36417 2:N:0:TGGTTCGA+TACCGTGT 
 # 6 A01346:18:H7KY5DRXY:2:2278:8956:36667 1:N:0:TGGTTCGA+TACCGTGT  A01346:18:H7KY5DRXY:2:2278:8956:36667 2:N:0:TGGTTCGA+TACCGTGT 
}
}
