% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat.R
\name{which_reformat.sh}
\alias{which_reformat.sh}
\title{Retrieve a path to the bbmap reformat.sh executable}
\usage{
which_reformat.sh(default = "/mnt/scgc_nfs/opt/common/bbmap/35.85/reformat.sh")
}
\arguments{
\item{default}{character, the default to use if \code{Sys.which()} doesn't find the application}
}
\value{
character, path to the executable
}
\description{
Retrieve a path to the bbmap reformat.sh executable
}
